split map from device stats/charts
