+++ netmanager/src/assets/css/manager-map.css
@@ -1,8 +1,7 @@
 .manager-map-container {
-    margin-top: 10px;
-    height: calc(100vh - 64px - 10px);
-    height: -moz-calc(100vh - 64px - 10px);
-    height: -webkit-calc(100vh - 64px - 10px);
+    height: calc(100vh - 64px);
+    height: -moz-calc(100vh - 64px);
+    height: -webkit-calc(100vh - 64px);
     position: relative;
 }
