+++ netmanager/src/assets/scss/device-management.sass
@@ -3,11 +3,16 @@
   margin: 0 auto
   position: relative
 
+.map-container-wrapper
+  width: 100%
+  margin: 0
+  position: relative
+
 
 .card-container
   width: 15%
   min-width: 150px
-  padding: 1px
+  padding: 5px
 
   &:hover
     cursor: pointer
@@ -115,7 +120,7 @@
 
 .map-container
   width: 100%
-  margin: 20px 0
+  margin: 0px
   position: relative
   height: calc(100vh - 300px)
